# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JOE Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get the current authenticated user's profile (JIT provisioned)
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update the current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      parameters:
        - name: role
          in: query
          schema:
            $ref: "#/components/schemas/UserRole"
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update a user's role (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found

  # ---------- Cohorts ----------
...
  /scholarship-applications:
    get:
      operationId: listScholarshipApplications
      tags: [scholarships]
      summary: List scholarship applications (admin)
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ScholarshipStatus"
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScholarshipApplication"
    post:
      operationId: createScholarshipApplication
      tags: [scholarships]
      summary: Apply for a scholarship as the current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScholarshipApplicationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScholarshipApplication"
...
  /scholarship-applications/{id}:
    get:
      operationId: getScholarshipApplication
      tags: [scholarships]
      summary: Get a scholarship application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScholarshipApplication"
        "404":
          description: Not found
  /scholarship-applications/{id}/review:
    patch:
      operationId: reviewScholarshipApplication
      tags: [scholarships]
      summary: Approve or reject a scholarship application (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScholarshipReviewInput"
      responses:
        "200":
          description: Reviewed application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScholarshipApplication"
        "404":
          description: Not found

  # ---------- Dashboards ----------
...
components:
  schemas:
...
    UserRole:
      type: string
      enum: [student, mentor, admin]
...
    ScholarshipStatus:
      type: string
      enum: [pending, approved, rejected]
    User:
      type: object
      required: [id, clerkId, email, name, role, createdAt]
      properties:
        id: { type: integer }
        clerkId: { type: string }
        email: { type: string }
        name: { type: string }
        role: { $ref: "#/components/schemas/UserRole" }
        avatarUrl: { type: ["string", "null"] }
        bio: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    UserUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        bio: { type: string }
        avatarUrl: { type: string }
    UserRoleUpdate:
      type: object
      required: [role]
      properties:
        role: { $ref: "#/components/schemas/UserRole" }
...
    ScholarshipApplication:
      type: object
      required: [id, applicantUserId, applicantName, applicantEmail, fullName, email, essay, status, appliedAt]
      properties:
        id: { type: integer }
        applicantUserId: { type: integer }
        applicantName: { type: string }
        applicantEmail: { type: string }
        cohortId: { type: ["integer", "null"] }
        fullName: { type: string }
        email: { type: string }
        essay: { type: string }
        status: { $ref: "#/components/schemas/ScholarshipStatus" }
        reviewerId: { type: ["integer", "null"] }
        reviewNotes: { type: ["string", "null"] }
        appliedAt: { type: string, format: date-time }
        reviewedAt: { type: ["string", "null"], format: date-time }
    ScholarshipApplicationInput:
      type: object
      required: [fullName, email, essay]
      properties:
        cohortId: { type: integer }
        fullName: { type: string, minLength: 1 }
        email: { type: string, minLength: 1 }
        essay: { type: string, minLength: 1 }
    ScholarshipReviewInput:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [approved, rejected] }
        reviewNotes: { type: string }
...